requirements.txt
tensorflow
streamlit
torch
torchvision
pillow
numpy
matplotlib
scipy
scikit-learn